Support __pairs and __ipairs (plus other tweaks)
Added PAIRS and IPAIRS constants to LuaValue
Merged pairs and ipairs into one class with a dynamic constructor, this
is similar to Lua's pairsmeta function.
Check for and call it's specific metamethod if it's available.

Add the colon after line numbers to error messages
Fix classes based on LibFunction giving "value" for their error message.